rename cpp/h files, update patch
